Move controller integration tests to integration dir
