+++ controller/validations_test.go
@@ -1,113 +1,81 @@
 package controller_test
 
-import (
-	"time"
+/*
+	Unit tests for the validations controller.
+	Integration tests using envtest can be found in the 'tests/integration/controller' directory.
 
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-	apinetworkingv1beta1 "istio.io/api/networking/v1beta1"
-	networkingv1beta1 "istio.io/client-go/pkg/apis/networking/v1beta1"
+	When should you write a unit test vs. writing an integration test? Favor integration tests over unit tests.
+	Write unit tests only when there's some specific state you need to test that you can't easily setup in an integration test.
+*/
+
+import (
+	"context"
+	"testing"
+
+	"github.com/stretchr/testify/require"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
+	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
+	"github.com/kiali/kiali/business"
+	"github.com/kiali/kiali/config"
+	"github.com/kiali/kiali/controller"
+	"github.com/kiali/kiali/kubernetes"
+	"github.com/kiali/kiali/kubernetes/cache"
+	"github.com/kiali/kiali/kubernetes/kubetest"
 	"github.com/kiali/kiali/models"
+	"github.com/kiali/kiali/store"
 )
 
-var _ = Describe("Validations controller", func() {
-	const (
-		timeout  = time.Second * 10
-		interval = time.Millisecond * 250
+type incrementFirstVersionStore[K comparable, V any] struct {
+	store.Store[K, V]
+	version uint
+}
+
+func (s *incrementFirstVersionStore[K, V]) Version() uint {
+	currentVersion := s.version
+	if currentVersion == 0 {
+		s.version++
+	}
+	return currentVersion
+}
+
+type incrementFirstVersionCache struct {
+	cache.KialiCache
+	validations *incrementFirstVersionStore[models.IstioValidationKey, *models.IstioValidation]
+}
+
+func (s *incrementFirstVersionCache) Validations() store.Store[models.IstioValidationKey, *models.IstioValidation] {
+	return s.validations
+}
+
+func newIncrementFirstVersionCache(cache cache.KialiCache) *incrementFirstVersionCache {
+	return &incrementFirstVersionCache{
+		KialiCache:  cache,
+		validations: &incrementFirstVersionStore[models.IstioValidationKey, *models.IstioValidation]{Store: cache.Validations()},
+	}
+}
+
+func TestValidationsFailsToUpdateWithOldCache(t *testing.T) {
+	require := require.New(t)
+	conf := config.NewConfig()
+	client := kubetest.NewFakeK8sClient(
+		&corev1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: "istio-system"}},
 	)
+	cache := newIncrementFirstVersionCache(business.SetupBusinessLayer(t, client, *conf))
+	k8sclients := map[string]kubernetes.ClientInterface{conf.KubernetesConfig.ClusterName: client}
+	namespace := business.NewNamespaceService(k8sclients, k8sclients, cache, *conf)
+	mesh := business.NewMeshService(k8sclients, cache, namespace, *conf)
+	layer := business.NewWithBackends(k8sclients, k8sclients, nil, nil)
+	validations := business.NewValidationsService(&layer.IstioConfig, cache, &mesh, &namespace, &layer.Svc, k8sclients, &layer.Workload)
+	reconciler := controller.NewValidationsReconciler([]string{conf.KubernetesConfig.ClusterName}, cache, &validations, 0)
 
-	Context("When validating a VirtualService", func() {
-		It("Should create validations in the kiali cache when a new VirtualService is created", func() {
-			Expect(kialiCache.Validations().Items()).Should(BeEmpty())
-
-			By("By creating a VirtualService")
-			istioSystemNamespace := &corev1.Namespace{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "istio-system",
-				},
-			}
-			Expect(k8sClient.Create(ctx, istioSystemNamespace)).Should(Succeed())
-			vs := &networkingv1beta1.VirtualService{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-vs",
-					Namespace: "default",
-				},
-				Spec: apinetworkingv1beta1.VirtualService{
-					Hosts:    []string{"test.com"},
-					Gateways: []string{"test-gateway"},
-				},
-			}
-			Expect(k8sClient.Create(ctx, vs)).Should(Succeed())
-
-			vsKey := types.NamespacedName{Name: vs.Name, Namespace: vs.Namespace}
-			createdVS := &networkingv1beta1.VirtualService{}
-
-			Eventually(func() bool {
-				err := k8sClient.Get(ctx, vsKey, createdVS)
-				return err == nil
-			}, timeout, interval).Should(BeTrue())
-			Expect(createdVS.Spec.Hosts).Should(Equal([]string{"test.com"}))
-
-			By("By checking that the validations are created in the kiali cache")
-			Eventually(func() bool {
-				validations := kialiCache.Validations().Items()
-				return len(validations) > 0
-			}, timeout, interval).Should(BeTrue())
-		})
-
-		It("Should update validations in the kiali cache when an existing VirtualService is updated", func() {
-			validationKey := models.IstioValidationKey{Name: "test-vs", Namespace: "default", ObjectType: "virtualservice"}
-			validation, err := kialiCache.Validations().Get(validationKey)
-			Expect(err).ShouldNot(HaveOccurred())
-			Expect(validation.Checks).Should(BeEmpty())
-			By("By updating a VirtualService")
-			vs := &networkingv1beta1.VirtualService{}
-			vsKey := types.NamespacedName{Name: "test-vs", Namespace: "default"}
-			Expect(k8sClient.Get(ctx, vsKey, vs)).Should(Succeed())
-			// Duplicate routes should be invalid.
-			vs.Spec.Http = []*apinetworkingv1beta1.HTTPRoute{
-				{
-					Route: []*apinetworkingv1beta1.HTTPRouteDestination{
-						{
-							Destination: &apinetworkingv1beta1.Destination{
-								Host: "test.com",
-							},
-						},
-						{
-							Destination: &apinetworkingv1beta1.Destination{
-								Host: "test.com",
-							},
-						},
-					},
-				},
-			}
-			Expect(k8sClient.Update(ctx, vs)).Should(Succeed())
-
-			By("By checking that the validations are then updated in the kiali cache")
-			Eventually(func() bool {
-				validation, err := kialiCache.Validations().Get(validationKey)
-				Expect(err).ShouldNot(HaveOccurred())
-				return len(validation.Checks) > 0
-			}, timeout, interval).Should(BeTrue())
-		})
-
-		It("Should delete validations in the kiali cache when the VirtualService is deleted", func() {
-			By("By deleting a VirtualService")
-			vs := &networkingv1beta1.VirtualService{}
-			vsKey := types.NamespacedName{Name: "test-vs", Namespace: "default"}
-			Expect(k8sClient.Get(ctx, vsKey, vs)).Should(Succeed())
-			Expect(k8sClient.Delete(ctx, vs)).Should(Succeed())
-
-			By("By checking that the validations are then deleted from the kiali cache")
-			Eventually(func() bool {
-				validationKey := models.IstioValidationKey{Name: "test-vs", Namespace: "default", ObjectType: "virtualservice"}
-				_, err := kialiCache.Validations().Get(validationKey)
-				return err != nil && len(kialiCache.Validations().Items()) == 0
-			}, timeout, interval).Should(BeTrue())
-		})
-	})
-})
+	// We want to test that the reconciler won't update the cache if the version has changed.
+	// Going to test this by having an implementation of the store which increments the version
+	// when you call it the first time. That way the next time it's called it will have a different
+	// version and the reconciler should not update the cache.
+	req := reconcile.Request{NamespacedName: types.NamespacedName{Name: "queue", Namespace: "queue"}}
+	_, err := reconciler.Reconcile(context.Background(), req)
+	require.Error(err)
+}
